+++ packages/app-portal/package.json
@@ -9,6 +9,7 @@
     "xstate:typegen": "xstate typegen 'src/**/*.ts?(x)'"
   },
   "peerDependencies": {
+    "@fuel-explorer/contract-ids": "workspace:*",
     "react": "18.2.0",
     "react-dom": "18.2.0"
   },
@@ -39,6 +40,7 @@
     "yup": "^1.2.0"
   },
   "devDependencies": {
+    "@fuel-explorer/contract-ids": "workspace:*",
     "@storybook/react": "^7.6.10",
     "@tanstack/react-query": "5.0.5",
     "@tanstack/react-query-devtools": "5.0.5",
