+++ packages/contract-ids/src/decodeMessageSentData.ts
@@ -1,0 +1,22 @@
+export const decodeMessageSentData = {
+  erc20Deposit: (data: `0x${string}`) => {
+    const pattern =
+      /^0x([A-f0-9]{64})([A-f0-9]{64})([A-f0-9]{64})([A-f0-9]{64})([A-f0-9]{64})([A-f0-9]{64})$/;
+    const match = data.match(pattern);
+    if (!match) {
+      return null;
+    }
+
+    const [, fuelTokenId, tokenAddress, , sender, to, amount] = match;
+
+    const parsed = {
+      fuelTokenId: `0x${fuelTokenId}`,
+      tokenAddress: `0x${tokenAddress}`,
+      sender: `0x${sender}`,
+      to: `0x${to}`,
+      amount,
+    };
+
+    return parsed;
+  },
+};
